[kernel] Build everything with -Wconversion [part 32]
Changes to be committed:
	modified:   include/tilck/kernel/ringbuf.h
	modified:   include/tilck/kernel/tasklet.h
	modified:   kernel/sched.c
	modified:   kernel/self_tests/se_data.c
	modified:   kernel/self_tests/se_data.h
	modified:   kernel/self_tests/se_kmalloc_perf.c
	modified:   kernel/self_tests/se_tasklet.c
	modified:   kernel/serial.c
	modified:   kernel/tasklet.c

+++ kernel/sched.c
@@ -168,7 +168,7 @@ void create_kernel_process(void)
 
    s_kernel_pi->ref_count = 1;
    s_kernel_ti->tid = kernel_pid;
-   s_kernel_ti->pid = kernel_pid;
+   s_kernel_ti->pid = (u16) kernel_pid;
 
    s_kernel_ti->pi = s_kernel_pi;
    init_task_lists(s_kernel_ti);
